double to int

+++ src/main/java/monopoly/field/Estate.java
@@ -3,16 +3,16 @@ package monopoly.field;
 public class Estate extends Property{
     private String color;
     private int houses;
-    private double[] rent;
+    private int[] rent;
 
-    public void Estate(String name, String color, double[] rent){
+    public void Estate(String name, String color, int[] rent){
         this.name = name;
         this.color = color;
         this.houses = 0;
         this.rent = rent;
     }
     @Override
-    public double getRent(){
+    public int getRent(){
         return rent[houses];
     }
