+++ src/main/java/monopoly/field/Trainstation.java
@@ -6,15 +6,12 @@ import java.util.stream.Collectors;
 public class Trainstation extends Property {
 
     @Override
-    public double getRent() {
+    public int getRent() {
         List<Property> properties = owner.getProperties();
         properties = properties.stream().filter(p -> p instanceof Trainstation).collect(Collectors.toList());
         int count = properties.size();
-        double price = count > 0 ? Math.pow(25,count-1) : 0;
+        int price = count > 0 ? Math.pow(25,count-1) : 0;
         return price;
     }
-
-
-
-
 }
+
